image_validation/requirements.txt
fastapi[standard]==0.115.12
psycopg2-binary==2.9.10
numpy==2.2.0
pillow==11.2.1
pydantic==2.11.3